upload/requirements.txt
absl-py==1.3.0
appdirs==1.4.4
apturl==0.5.2
asgiref==3.5.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.2.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.2.0
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cv==1.0.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
Deprecated==1.2.13
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==4.0.5
django-braces==1.15.0
django-comments-dab==2.8.0
django-crispy-forms==1.14.0
django-gravatar2==1.4.4
django-model-utils==4.3.1
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
django-star-ratings==0.9.2
django-utils-six==2.0
django-widget-tweaks==1.4.12
docopt==0.6.2
duplicity==0.8.21
fasteners==0.14.1
filelock==3.6.0
flatbuffers==22.10.26
fonttools==4.29.1
fs==2.4.12
future==0.16.0
gast==0.4.0
gdown==4.6.0
glob2==0.7
google-auth==2.14.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
imageio==2.22.3
importlib-metadata==4.6.4
ipython==7.31.1
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.2
keras==2.10.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==14.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
netifaces==0.11.0
numpy==1.21.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.0
olefile==0.46
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
paramiko==2.9.3
parso==0.8.1
pbr==5.9.0
pexpect==4.8.0
phonenumbers==8.12.51
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.2
ply==3.11
prompt-toolkit==3.0.28
protobuf==3.12.4
protonvpn-cli==2.2.11
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-decouple==3.6
pythondialog==3.5.3
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==4.3.4
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
ssh-import-id==5.11
stevedore==3.5.0
swapper==1.3.0
sympy==1.9
systemd-python==234
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
traitlets==5.1.1
typing_extensions==4.3.0
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.14.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wget==3.2
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0
